# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Mirrors https://thesimpsonsapi.com/api. JSON responses are proxied from the upstream service; pagination next/prev URLs reference the upstream host.",
        "title": "The Simpsons API (proxied)",
        "termsOfService": "https://thesimpsonsapi.com/",
        "contact": {
            "name": "API support",
            "url": "https://thesimpsonsapi.com/"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api": {
            "get": {
                "description": "Returns links to the characters, episodes, and locations collection endpoints.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meta"
                ],
                "summary": "API index",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIIndex"
                        }
                    }
                }
            }
        },
        "/api/characters": {
            "get": {
                "description": "Paginated list of characters (20 per page). Use the page query parameter.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "characters"
                ],
                "summary": "List characters",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (1-based)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PaginatedCharacters"
                        }
                    }
                }
            }
        },
        "/api/characters/{id}": {
            "get": {
                "description": "Full character record including description and optional nested first-appearance episode.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "characters"
                ],
                "summary": "Get character by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Character ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CharacterDetail"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/episodes": {
            "get": {
                "description": "Paginated list of episodes (20 per page).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "episodes"
                ],
                "summary": "List episodes",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (1-based)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PaginatedEpisodes"
                        }
                    }
                }
            }
        },
        "/api/episodes/{id}": {
            "get": {
                "description": "Episode detail including long description.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "episodes"
                ],
                "summary": "Get episode by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Episode ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.EpisodeDetail"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/locations": {
            "get": {
                "description": "Paginated list of Springfield locations (20 per page).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "locations"
                ],
                "summary": "List locations",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (1-based)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PaginatedLocations"
                        }
                    }
                }
            }
        },
        "/api/locations/{id}": {
            "get": {
                "description": "Location detail including description and optional nested appearances.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "locations"
                ],
                "summary": "Get location by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Location ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.LocationDetail"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.APIIndex": {
            "type": "object",
            "properties": {
                "characters": {
                    "type": "string",
                    "example": "https://thesimpsonsapi.com/api/characters"
                },
                "episodes": {
                    "type": "string",
                    "example": "https://thesimpsonsapi.com/api/episodes"
                },
                "locations": {
                    "type": "string",
                    "example": "https://thesimpsonsapi.com/api/locations"
                }
            }
        },
        "models.CharacterDetail": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "birthdate": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "first_appearance_ep": {
                    "$ref": "#/definitions/models.EpisodeDetail"
                },
                "first_appearance_ep_id": {
                    "type": "integer"
                },
                "first_appearance_sh_id": {
                    "type": "integer"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "occupation": {
                    "type": "string"
                },
                "phrases": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "portrait_path": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "models.CharacterListItem": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "birthdate": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "occupation": {
                    "type": "string"
                },
                "phrases": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "portrait_path": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "models.EpisodeBrief": {
            "type": "object",
            "properties": {
                "airdate": {
                    "type": "string"
                },
                "episode_number": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "image_path": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "season": {
                    "type": "integer"
                },
                "synopsis": {
                    "type": "string"
                }
            }
        },
        "models.EpisodeDetail": {
            "type": "object",
            "properties": {
                "airdate": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "episode_number": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "image_path": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "season": {
                    "type": "integer"
                },
                "synopsis": {
                    "type": "string"
                }
            }
        },
        "models.LocationBrief": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "image_path": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "town": {
                    "type": "string"
                },
                "use": {
                    "type": "string"
                }
            }
        },
        "models.LocationDetail": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "first_appearance_ep": {
                    "$ref": "#/definitions/models.EpisodeDetail"
                },
                "first_appearance_ep_id": {
                    "type": "integer"
                },
                "first_appearance_sh": {
                    "type": "object"
                },
                "first_appearance_sh_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "image_path": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "town": {
                    "type": "string"
                },
                "use": {
                    "type": "string"
                }
            }
        },
        "models.PaginatedCharacters": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "next": {
                    "type": "string"
                },
                "pages": {
                    "type": "integer"
                },
                "prev": {
                    "type": "string"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CharacterListItem"
                    }
                }
            }
        },
        "models.PaginatedEpisodes": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "next": {
                    "type": "string"
                },
                "pages": {
                    "type": "integer"
                },
                "prev": {
                    "type": "string"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.EpisodeBrief"
                    }
                }
            }
        },
        "models.PaginatedLocations": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "next": {
                    "type": "string"
                },
                "pages": {
                    "type": "integer"
                },
                "prev": {
                    "type": "string"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.LocationBrief"
                    }
                }
            }
        }
    }
}